requirements.txt
django
djangorestframework
drf_yasg==1.21.7
django-cors-headers
django-jalali
jdatetime
mysqlclient
django-cors-headers